x64 compat fixes

+++ ddclexer/src/ddc/lexer/exceptions.d
@@ -5,12 +5,12 @@ import std.conv;
 class ParserException : Exception {
     string _msg;
     string _filename;
-    int _line;
-    int _pos;
+    size_t _line;
+    size_t _pos;
 
-    public @property int line() { return _line; }
+    public @property size_t line() { return _line; }
 
-    this(string msg, string filename, int line, int pos) {
+    this(string msg, string filename, size_t line, size_t pos) {
         super(msg ~ " at " ~ filename ~ " line " ~ to!string(line) ~ " column " ~ to!string(pos));
         _msg = msg;
         _filename = filename;
@@ -20,13 +20,13 @@ class ParserException : Exception {
 }
 
 class LexerException : ParserException {
-    this(string msg, string filename, int line, int pos) {
+    this(string msg, string filename, size_t line, size_t pos) {
         super(msg, filename, line, pos);
     }
 }
 
 class SourceEncodingException : LexerException {
-    this(string msg, string filename, int line, int pos) {
+    this(string msg, string filename, size_t line, size_t pos) {
         super(msg, filename, line, pos);
     }
 }
